Added my website Etherpunks.com to whitelist
My website is rapidly being built.  It uses PersonalSign to allow a user to 'login using metamask' then it imports their digital assets such as their Cryptopunks and (soon) decentraland plots.  It acts as a social network for these digital identities.  When the tokens are given to other people, then those people assume control over the profile so control of the account is tied to the token, not tied to a person.